manage student completed

+++ back-end/routes/api.php
@@ -3,10 +3,14 @@
 use App\Http\Controllers\AdminController;
 use App\Http\Controllers\AuthController;
 use App\Http\Controllers\ClassController;
+use App\Http\Controllers\StudentController;
 use App\Http\Middleware\CheckRole;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\TeacherController;
+use App\Http\Controllers\UserController;
+use App\Http\Middleware\CheckAuthentication;
+use Tymon\JWTAuth\Facades\JWTAuth;
 
 Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
     return $request->user();
@@ -16,7 +20,6 @@ Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
 Route::prefix("auth")->group(function() {
     Route::post("/checkUserLogin", [AuthController::class, "checkUserLogin"]); // check withen user credentials is valid
 });
-//
 
 
 Route::prefix("admin")->middleware(CheckRole::class . ":admin")->group(function(){
@@ -26,15 +29,31 @@ Route::prefix("admin")->middleware(CheckRole::class . ":admin")->group(function(
     Route::delete("/deleteAdmin/{id}", [AdminController::class, "deleteAdmin"]);
 });
 
-Route::prefix("class")->group(function(){
-    Route::get("/getClasses", [ClassController::class ,"getClasses"]);
-    Route::post("/createClass", [ClassController::class ,"createClass"]);
-    Route::delete("/deleteClass/{id}", [ClassController::class, "deleteClass"]);
+
+Route::prefix("user")->middleware(CheckRole::class . ":admin")->group(function(){
+    Route::put("/updateUserData/{id}", [UserController::class, 'updateUserCredentials']);
 });
 
+
+Route::prefix("student")->group(function(){
+    Route::get("/getStudents", [StudentController::class, "getStudents"])->middleware(CheckAuthentication::class);
+    Route::get("/getStudent/{id}", [StudentController::class, "getStudent"])->middleware(CheckAuthentication::class);
+    Route::post("/createStudent", [StudentController::class, "createStudent"])->middleware(CheckRole::class . ':admin');
+    Route::put("/updateStudentData", [StudentController::class, "updateStudent"])->middleware(CheckAuthentication::class);
+    Route::delete("/deleteStudent/{id}", [StudentController::class, "deleteStudent"])->middleware(CheckRole::class . ':admin');
+});
+
+
 Route::prefix("teacher")->group(function(){
-    Route::get("/getTeachers",[TeacherController::class,"getTeachers"]);
-    Route::post("/createTeacher",[TeacherController::class,"createTeacher"]);
-    Route::put("/updateTeacher/{id}",[TeacherController::class,"updateTeacher"]);
-    Route::delete("/deleteTeacher/{id}",[TeacherController::class,"deleteTeacher"]);
+    Route::get("/getTeachers",[TeacherController::class,"getTeachers"])->middleware(CheckRole::class . ":admin");
+    Route::post("/createTeacher",[TeacherController::class,"createTeacher"])->middleware(CheckRole::class . ":admin");
+    Route::put("/updateTeacher/{id}",[TeacherController::class,"updateTeacher"])->middleware(CheckAuthentication::class);
+    Route::delete("/deleteTeacher/{id}",[TeacherController::class,"deleteTeacher"])->middleware(CheckRole::class . ":admin");
+});
+
+
+Route::prefix("class")->group(function(){
+    Route::get("/getClasses", [ClassController::class ,"getClasses"])->middleware(CheckAuthentication::class);
+    Route::post("/createClass", [ClassController::class ,"createClass"])->middleware(CheckRole::class . ":admin");
+    Route::delete("/deleteClass/{id}", [ClassController::class, "deleteClass"])->middleware(CheckRole::class . ":admin");
 });
